backend/requirements/common.txt
django==2.2.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.0.2
dj-database-url==0.5.0
psycopg2==2.8.2
whitenoise==4.1.2
